==> 00_DL_fuse_biodiscover_metadata.ipynb <==
# Using Python 3.12.1 (local env: HeatNSalt)

import pandas as pd
import os

# Fuse BioDiscover image data files 
path_image = 'raw/raw_image_data'
image_files = os.listdir(path_image)
if '.DS_Store' in image_files:
    image_files.remove('.DS_Store')

complete_image_df = pd.read_csv(path_image+'/'+image_files[0],encoding='ISO-8859-1', sep=';') # read first csv file
for file in image_files[1:]: # read remaining csv files and add to first dataframe
    new_image_df = pd.read_csv(path_image+'/'+file,encoding='ISO-8859-1',sep=';')
    complete_image_df = pd.merge(complete_image_df, new_image_df, how='outer')

complete_image_df.to_csv('outputs/image_data_complete.csv', index=False)

# Fuse BioDiscover setup data files
path_setup = 'raw/raw_setup_data'
setup_files_csv = [file for file in os.listdir(path_setup) if file.endswith('.csv')]
setup_files_xlsx = [file for file in os.listdir(path_setup) if file.endswith('.xlsx')]


complete_setup_df = pd.read_csv(path_setup+'/'+setup_files_csv[0],encoding='ISO-8859-1', sep=';') # read first csv file
complete_setup_df = complete_setup_df.drop(['Longitude'], axis=1)
complete_setup_df = complete_setup_df.drop(['Latitude'], axis=1)

for file in setup_files_csv[1:]: # read remaining csv files and add to first dataframe
    new_setup_df = pd.read_csv(path_setup+'/'+file,encoding='ISO-8859-1',sep=';')
    new_setup_df = new_setup_df.drop(['Longitude'], axis=1)
    new_setup_df = new_setup_df.drop(['Latitude'], axis=1)
    complete_setup_df = pd.merge(complete_setup_df, new_setup_df, how='outer')

complete_setup_df['Date (DD/MM/YYYY HH:MM)'] = pd.to_datetime(complete_setup_df['Date (DD/MM/YYYY HH:MM)'], format='%d.%m.%Y %H:%M')

for file in setup_files_xlsx: # read all xlsx files and add to first dataframe
    new_setup_df = pd.read_excel(path_setup+'/'+file)
    new_setup_df = new_setup_df.drop(['Longitude'], axis=1)
    new_setup_df = new_setup_df.drop(['Latitude'], axis=1)
    new_setup_df['Date (DD/MM/YYYY HH:MM)'] = pd.to_datetime(new_setup_df['Date (DD/MM/YYYY HH:MM)'], format='%d.%m.%Y %H:%M')
    complete_setup_df = pd.merge(complete_setup_df, new_setup_df, how='outer')

complete_setup_df = complete_setup_df.rename(columns={'Latitude': 'Tube_weight_empty', 'Longitude': 'Tube_weight_with_dried_specimen', 'Sex': 'Order'})

complete_setup_df.to_csv('outputs/setup_data_complete.csv', index=False)
