==> 02_DL_process_second_OTU_table.ipynb <==
# Using Python 3.12.1 (local env: HeatNSalt)

# Process the OTU table of the repeated plate 6, starting with BOLD results. 

import pandas as pd
import numpy as np
import tiers

BOLD6_df = pd.read_excel('raw/ExStreamEPT6_apscale_OTUs_identification_result.xlsx', engine='openpyxl')

BOLD6_df.iloc[:,:14]

# Separate metadata / taxa info from read_counts
otu_info = BOLD6_df.iloc[:,:14]
read_matrix = BOLD6_df.set_index('id').iloc[:,14:]

# Extract OTU number for sorting
otu_info = otu_info.assign(idno=otu_info['id'].map(lambda x: x.split('_')[1]))
otu_info['idno'] = otu_info['idno'].astype(int)
otu_info = otu_info.sort_values('idno').reset_index(drop=True)

# Create a taxon hierarchy table
# Separate taxon information and add Kingdom value
otu_taxa_table = otu_info.iloc[:,:7].set_index('id')
otu_taxa_table.insert(loc=0, column='Kingdom', value='Animalia')

# set column names to lowercase
otu_taxa_table.columns = otu_taxa_table.columns.map(lambda x: x.lower())
otu_taxa_table = otu_taxa_table.assign(type=np.nan)
otu_taxa_table = otu_taxa_table.assign(label=tiers.get_leaves(otu_taxa_table))

# Merge to final otu_info table
otu_info = otu_taxa_table.merge(otu_info, left_index=True, right_on='id')

# Drop the old hierarchy before merging
otu_info = otu_info.drop(['Phylum', 'Class', 'Order', 'Family', 'Genus', 'species'],axis=1)

# Duplicates are removed from otu_taxa_table
otu_taxa_table = otu_taxa_table.drop_duplicates(keep='first').sort_values(otu_taxa_table.columns.tolist()).reset_index(drop=True)

read_matrix


# Handle read matrix
otu_sums = read_matrix.sum(axis=1)
any(otu_sums != 0)
# if True, Each OTU has some reads

# Function to get ordered list of non-zero values with indices
def non_zero_values(df):
    result = {}
    for col in df.columns:
        non_zero_values = df[col][df[col] != 0]  # Filter non-zero values
        if non_zero_values.empty:
            result[col] = []
            continue
        ordered_values = non_zero_values.sort_values()  # Sort by value
        result[col] = list(zip(ordered_values.index.astype(str), ordered_values))  # Store indices and values as list of tuples
    return result

ordered_non_zero_values = non_zero_values(read_matrix)

def replace_string_in_tuple(tup, df):
    id_value, reads = tup
    row = df.loc[df['id'] == id_value]

    # Check if the ID in the tuple matches with the ID in the DataFrame
    if not row.empty:
        # Find the corresponding label value from the DataFrame
        corresponding_label = row['label'].iloc[0]
        corresponding_species = row['Species'].iloc[0]
        corresponding_genus = row['genus'].iloc[0]
        corresponding_family = row['family'].iloc[0]
        corresponding_order = row['order'].iloc[0]
        # Replace the label value in the tuple
        return (id_value, corresponding_order, corresponding_family, corresponding_genus, corresponding_species, corresponding_label, reads)
    else:
        return tup  # If no matching ID is found, return the original tuple
    
species_dict = {}
count = 0
for sample, hits in ordered_non_zero_values.items():
    for hit in hits:
        species_dict[count] = [sample, replace_string_in_tuple(hit, otu_info)]
        count += 1

Megabarcoding_df = pd.DataFrame.from_dict(species_dict, orient='index')
Megabarcoding_df[['OTU', 'Order', 'Family', 'Genus', 'Species', 'Label', 'read count']] = pd.DataFrame(Megabarcoding_df[1].tolist(), index=Megabarcoding_df.index)
Megabarcoding_df = Megabarcoding_df.drop(columns=[1])
Megabarcoding_df = Megabarcoding_df.sort_values(by=[0, 'read count'], ascending=[True, False])

Megabarcoding_df = Megabarcoding_df.rename(columns={0: 'Sample'})
Megabarcoding_max_df = Megabarcoding_df.drop_duplicates(subset='Sample')

with pd.ExcelWriter('outputs/EPT_plate6_OTU_reads.xlsx') as writer:
   
    # use to_excel function and specify the sheet_name and index 
    # to store the dataframe in specified sheet
    Megabarcoding_max_df.to_excel(writer, sheet_name='Top Hits', index=False)
    Megabarcoding_df.to_excel(writer, sheet_name='All reads', index=False)


Megabarcoding_df_big = pd.read_excel('outputs/EPT_OTU_reads.xlsx')
merged = Megabarcoding_max_df.merge(Megabarcoding_df_big, on='Sample', suffixes=('_df1', '_df2'))
merged['Label_equal'] = merged['Label_df1'] == merged['Label_df2']

merged

differences = merged[merged['Label_df1'] != merged['Label_df2']]
differences
