==> 04_DL_decide_DNA_label.ipynb <==
# Using Python 3.12.1 (local env: HeatNSalt)

import pandas as pd
import numpy as np

## The images were sorted manually by moving the 'bad' images to a new folder via Windows explorer. This folder was scanned:
# bad_img_df = pd.DataFrame(os.listdir('E:\Biodiscover_Vault\PMR_ExStream22_Main_EPT\Filtered images'), columns=['bad_img'])
# bad_img_df.to_excel('EPT_filtered_images.xlsx')

bad_img_df = pd.read_excel('raw/EPT_filtered_images.xlsx')
IMG_df = pd.read_csv('outputs/image_data_with_DNA.csv')
All_IMGs_df = pd.read_csv('outputs/image_data_complete.csv')

print(f'Total number of images collected: {len(All_IMGs_df.index)}')
print(f'Number of images after quality filtering: {len(All_IMGs_df.index) - len(bad_img_df.index)}')

cleaning_list = IMG_df['Image File Name'].isin(bad_img_df['bad_img'])
image_data_with_DNA_clean = IMG_df[~cleaning_list]

image_data_with_DNA_clean.to_csv('outputs/image_data_with_DNA_clean.csv')

# Delete DNA info if reads below 1000 + COUNT
# Rename image_data_with_DNA_clean to df for handling reasons:
df = image_data_with_DNA_clean

# Identify columns that end with '_DNA'
data_columns = [col for col in df.columns if '_DNA' not in col]
base_df = df[data_columns]
dna_columns = [col for col in df.columns if col.endswith('_DNA')]
repeat_columns = [col for col in df.columns if col.endswith('_DNA_repeat')]

# Set values to NaN where 'read count_DNA' or 'read count_DNA_repeat' is below 1000

print(f"Number of specimen with <1000 reads in first samples: {df.loc[df['read count_DNA'] < 1000, 'Specimen ID'].nunique()}")
not_in_final_df = {}
not_in_final_df['delete_in_first'] = df.loc[df['read count_DNA'] < 1000, 'Specimen ID'].unique()
print(f"Number of specimen with <1000 reads in repeated samples: {df.loc[df['read count_DNA_repeat'] < 1000, 'Specimen ID'].nunique()}")
not_in_final_df['delete_in_repeat'] = df.loc[df['read count_DNA_repeat'] < 1000, 'Specimen ID'].unique()

df.loc[df['read count_DNA'] < 1000, dna_columns] = np.nan
df.loc[df['read count_DNA_repeat'] < 1000, repeat_columns] = np.nan


# Delete DNA info if assigned to wrong order for DNA and DNA_repeat
non_match_mask = df['Order_DNA'].notna() & (df['Order_DNA'] != df['Order_Morpho'])
repeat_non_match_mask = df['Order_DNA_repeat'].notna() & (df['Order_DNA_repeat'] != df['Order_Morpho'])

print(f"Number of specimen where Morpho_Order =/= DNA_Order: {df.loc[non_match_mask, 'Specimen ID'].nunique()}")
not_in_final_df['Morpho_Order =/= DNA_Order'] = df.loc[non_match_mask, 'Specimen ID'].unique()
print(f"Number of specimen where Morpho_Order =/= DNA_repeat_Order: {df.loc[repeat_non_match_mask, 'Specimen ID'].nunique()}")
not_in_final_df['Morpho_Order =/= DNA_repeat_Order'] = df.loc[repeat_non_match_mask, 'Specimen ID'].unique()

df.loc[non_match_mask, dna_columns] = np.nan
df.loc[repeat_non_match_mask, repeat_columns] = np.nan

def determine_final_label(row):
    if pd.isna(row['read count_DNA']) and pd.isna(row['read count_DNA_repeat']):
        return pd.Series([np.nan] * len(dna_columns), index=dna_columns)  # Both values are NaN
    elif pd.isna(row['read count_DNA']):
        # Only 'read count_DNA' is NaN, return the '_repeat' values and rename columns
        repeat_values = row[repeat_columns].copy()
        repeat_values.index = [col.replace('_repeat', '') for col in repeat_values.index]
        return repeat_values
    elif pd.isna(row['read count_DNA_repeat']):
        return row[dna_columns]  # Only 'read count_DNA_repeat' is NaN
    else:
        # Neither value is NaN, return the row with higher 'read count' and rename columns if needed
        if row['read count_DNA'] > row['read count_DNA_repeat']:
            return row[dna_columns]
        else:
            repeat_values = row[repeat_columns].copy()
            repeat_values.index = [col.replace('_repeat', '') for col in repeat_values.index]
            return repeat_values

# Decide if the first run or if the repeated run is chosen based on read count
DNA_df = df.apply(determine_final_label, axis = 1)
DNA_df = DNA_df.reindex(columns=['OTU_DNA', 'Order_DNA', 'Family_DNA', 'Genus_DNA', 'Species_DNA', 'Label_DNA', 'read count_DNA'])
EPT_DL_df = base_df.join(DNA_df)

print(f'Number of specimen remaining: {EPT_DL_df['Specimen ID'].nunique()}')
print(f'Corresponding number of images: {EPT_DL_df['Specimen ID'].count()}')
EPT_DL_species_df = EPT_DL_df.dropna(subset=['Species_DNA'])
print(f'Number of OTUs (/classes): {EPT_DL_species_df['OTU_DNA'].nunique()}')
print(f'Number of specimen on species level: {EPT_DL_species_df['Specimen ID'].nunique()}')
unique_counts = EPT_DL_species_df.groupby('Species_DNA')['Specimen ID'].nunique().sort_values(ascending=False)
print(f'Corresponding number of images: {EPT_DL_species_df['Specimen ID'].count()}')

pd.DataFrame(unique_counts)

not_in_final_df

list(set(not_in_final_df['delete_in_first']).intersection(not_in_final_df['delete_in_repeat']))

IMG_df[IMG_df['Specimen ID']== '3_B12']

subset = IMG_df[IMG_df['Specimen ID'].isin(not_in_final_df['delete_in_first'])]
filtered = subset[subset['Order_DNA_repeat'].isna()]

filtered['Specimen ID'].unique()


# Remove species with less than 5 specimen:
print('Specimen that will be removed:')
print(EPT_DL_species_df[EPT_DL_species_df['Species_DNA'].isin(unique_counts[unique_counts < 5].index)]['Specimen ID'].unique())
EPT_14_DL_df =  EPT_DL_species_df[~EPT_DL_species_df['Species_DNA'].isin(unique_counts[unique_counts < 5].index)]

print(f'Number of specimen that will be used for deep learning: {EPT_14_DL_df['Specimen ID'].nunique()}')
print(f'Number of species (/classes): {EPT_14_DL_df['Species_DNA'].nunique()}')
print(f'Corresponding number of images: {EPT_14_DL_df['Specimen ID'].count()}')

EPT_14_DL_df.to_csv('outputs/EPT-14-DL.csv', index=False)

unique_counts_dl = EPT_14_DL_df.groupby('Species_DNA')['Specimen ID'].nunique().sort_values(ascending=False)

pd.DataFrame(unique_counts_dl)

Compare failed samples with rest of the samples

# Values in IMG_df['Specimen ID'] but not in All_IMGs_df['Sample Name/Number']
All_IMGs_df['Sample Name/Number'] = All_IMGs_df['Sample Name/Number'].apply(lambda x: x.rsplit('_', 1)[0])
All_IMGs_df[~All_IMGs_df['Sample Name/Number'].isin(EPT_14_DL_df['Specimen ID'])].dropna(axis=1).describe().round(2)

All_IMGs_df[All_IMGs_df['Sample Name/Number'].isin(EPT_14_DL_df['Specimen ID'])].dropna(axis=1).describe().round(2)

IMG_df[~IMG_df['Specimen ID'].isin(EPT_14_DL_df['Specimen ID'])].dropna(axis=1).describe().round(2)

IMG_df[IMG_df['Specimen ID'].isin(EPT_14_DL_df['Specimen ID'])].dropna(axis=1).describe().round(2)
