==> 06_DL_create_checklist.ipynb <==
# Using Python 3.12.1 (local env: HeatNSalt)

## In a first round of training different learning rates and different number of images used as input were tested.
## The images of the specimens that were most commonly identified incorrectly were checked by an taxonomic expert (JE).
import pandas as pd
import random
random.seed(42)

ckt = pd.read_excel('Wrong Identifications with names_JE_2.xlsx')
filtered_df = ckt[(ckt['morphological ID'].isna()) & (ckt['count'] < 5)]
all_df = pd.read_csv('outputs/EPT-14-DL.csv')
rare_species = ['Nemoura cinerea', 'Baetis vernus', 'Amphinemura standfussi', 'Mystacides azureus', 'Anabolia nervosa', 'Halesus radiatus']

print(f'Total number of specimen: {all_df['Specimen ID'].nunique()}')
print(f'Number of checked specimen: {filtered_df['individual'].nunique()}')

specimen = all_df['Specimen ID'].unique().tolist()
not_this =  filtered_df['individual'].unique().tolist()

new_ls = [x for x in specimen if x not in not_this]

random_ls = random.sample(new_ls, k=100)
random_df = all_df[all_df['Specimen ID'].isin(random_ls)]
selected_df = all_df[all_df['Species_DNA'].isin(rare_species)]

to_check_df = pd.concat([random_df, selected_df], ignore_index=True)
to_check_df = to_check_df.drop_duplicates(subset='Specimen ID', keep='first')
to_check_df.to_excel('outputs/EPT-14-Checklist.xlsx', index=False)

pd.DataFrame(to_check_df.groupby('Species_DNA')['Specimen ID'].nunique().sort_values(ascending=False))
